+++ moveit_ros/hybrid_planning/CMakeLists.txt
@@ -14,6 +14,7 @@ find_package(moveit_msgs REQUIRED)
 find_package(moveit_ros_planning REQUIRED)
 find_package(moveit_core REQUIRED)
 find_package(std_srvs REQUIRED)
+find_package(std_msgs REQUIRED)
 find_package(tf2_ros REQUIRED)
 find_package(trajectory_msgs REQUIRED)
 find_package(control_toolbox REQUIRED)
@@ -30,6 +31,7 @@ set(LIBRARIES
   single_plan_execution_plugin
   decelerate_before_collision_plugin
   replan_invalidated_trajectory_plugin
+  forward_trajectory_plugin
 )
 
 set(THIS_PACKAGE_INCLUDE_DEPENDS
@@ -39,6 +41,7 @@ set(THIS_PACKAGE_INCLUDE_DEPENDS
   moveit_core
   moveit_ros_planning
   moveit_msgs
+  std_msgs
   std_srvs
   tf2_ros
   trajectory_msgs
@@ -47,20 +50,17 @@ set(THIS_PACKAGE_INCLUDE_DEPENDS
 
 set(THIS_PACKAGE_INCLUDE_DIRS
   global_planner/include
-  local_planner/include
-  hybrid_planning_manager/include
-  planner_logic_plugins/include
-  trajectory_operator_plugins/include
-  local_constraint_solver_plugins/include
+  hybrid_planning_manager/hybrid_planning_manager_component/include
+  hybrid_planning_manager/planner_logic_plugins/include
+  local_planner/local_planner_component/include
+  local_planner/trajectory_operator_plugins/include
+  local_planner/local_constraint_solver_plugins/include
 )
 include_directories(${THIS_PACKAGE_INCLUDE_DIRS})
 
 add_subdirectory(hybrid_planning_manager)
 add_subdirectory(global_planner)
 add_subdirectory(local_planner)
-add_subdirectory(planner_logic_plugins)
-add_subdirectory(trajectory_operator_plugins)
-add_subdirectory(local_constraint_solver_plugins)
 add_subdirectory(test)
 
 rclcpp_components_register_nodes(moveit_hybrid_planning_manager "moveit_hybrid_planning::HybridPlanningManager")
@@ -70,8 +70,7 @@ rclcpp_components_register_nodes(moveit_local_planner_component "moveit_hybrid_p
 install(TARGETS ${LIBRARIES} hybrid_planning_test_node
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin
-  DESTINATION lib/${PROJECT_NAME})
+  RUNTIME DESTINATION lib/${PROJECT_NAME})
 
 install(DIRECTORY ${THIS_PACKAGE_INCLUDE_DIRS} DESTINATION include)
 
@@ -82,6 +81,7 @@ pluginlib_export_plugin_description_file(moveit_hybrid_planning single_plan_exec
 pluginlib_export_plugin_description_file(moveit_hybrid_planning next_waypoint_sampler_plugin.xml)
 pluginlib_export_plugin_description_file(moveit_hybrid_planning decelerate_before_collision_plugin.xml)
 pluginlib_export_plugin_description_file(moveit_hybrid_planning replan_invalidated_trajectory_plugin.xml)
+pluginlib_export_plugin_description_file(moveit_hybrid_planning forward_trajectory_plugin.xml)
 
 ament_export_include_directories(include)
 ament_export_libraries(${LIBRARIES})
