+++ moveit_ros/hybrid_planning/local_planner/local_constraint_solver_plugins/CMakeLists.txt
@@ -1,0 +1,13 @@
+add_library(decelerate_before_collision_plugin SHARED
+  src/decelerate_before_collision.cpp
+)
+set_target_properties(decelerate_before_collision_plugin PROPERTIES VERSION "${${PROJECT_NAME}_VERSION}")
+ament_target_dependencies(decelerate_before_collision_plugin ${THIS_PACKAGE_INCLUDE_DEPENDS})
+target_link_libraries(decelerate_before_collision_plugin local_constraint_solver_interface)
+
+add_library(forward_trajectory_plugin SHARED
+  src/forward_trajectory.cpp
+)
+set_target_properties(forward_trajectory_plugin PROPERTIES VERSION "${${PROJECT_NAME}_VERSION}")
+ament_target_dependencies(forward_trajectory_plugin ${THIS_PACKAGE_INCLUDE_DEPENDS})
+target_link_libraries(forward_trajectory_plugin local_constraint_solver_interface)
