+++ moveit_ros/hybrid_planning/package.xml
@@ -23,6 +23,7 @@
   <depend>moveit_core</depend>
   <depend>moveit_ros_planning</depend>
   <depend>tf2_ros</depend>
+  <depend>std_msgs</depend>
   <depend>std_srvs</depend>
   <depend>trajectory_msgs</depend>
   <depend>pluginlib</depend>
@@ -37,6 +38,7 @@
     <moveit_hybrid_planning plugin="${prefix}/next_waypoint_sampler_plugin.xml"/>
     <moveit_hybrid_planning plugin="${prefix}/replan_invalidated_trajectory_plugin.xml"/>
     <moveit_hybrid_planning plugin="${prefix}/single_plan_execution_plugin.xml"/>
+    <moveit_hybrid_planning plugin="${prefix}/forward_trajectory_plugin.xml"/>
   </export>
 
 </package>
