+++ moveit_ros/hybrid_planning/test/config/hybrid_planning_manager.yaml
@@ -1,3 +1,1 @@
-
-# planner_logic_plugin_name: "moveit_hybrid_planning/SinglePlanExecution"
 planner_logic_plugin_name: "moveit_hybrid_planning/ReplanInvalidatedTrajectory"
